requirements_ai.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.12.0
python-dotenv==1.0.0
